Created the total and implemented the hide/unhide functionality
I created the total html and implemented hiding and unhiding of said element. The element will be hidden when there are no items in the cart and will be unhidden when there are items in the cart. I also commented out the first line of the cart.js file because it was breaking the code.